flow: FNV-1a FlowKeyHash with uint64 accumulator (32-bit-safe)
The mipsel 32-bit cross build failed with -Werror=overflow on the FNV offset
basis (14695981039346656037ULL) being assigned to a 32-bit size_t. Keep the
accumulator uint64_t and truncate only on return - no constant-overflow at
init, portable to 32-bit targets and macOS without a vendored hash header.
(vendored xxhash experiment dropped - it failed to compile on the dev macOS
toolchain, so FNV-1a over the packed 46B key stays.)

+++ src/flow.hpp
@@ -55,17 +55,19 @@ inline bool operator!=(const FlowKey &a, const FlowKey &b)
 	return !(a == b);
 }
 
-/* FNV-1a over the packed key bytes; hash functor for unordered_set<FlowKey>. */
+/* FNV-1a over the packed key bytes; hash functor for unordered_set<FlowKey>.
+ * The accumulator stays uint64_t; only the final value is truncated to
+ * size_t, so this compiles warning-clean on 32-bit targets (mipsel) as well. */
 struct FlowKeyHash {
 	size_t operator()(const FlowKey &k) const
 	{
 		const uint8_t *b = reinterpret_cast<const uint8_t *>(&k);
-		size_t h = 14695981039346656037ULL; /* FNV offset basis */
+		uint64_t h = 14695981039346656037ULL; /* FNV offset basis */
 		for (size_t i = 0; i < sizeof(FlowKey); i++) {
 			h ^= b[i];
 			h *= 1099511628211ULL; /* FNV prime */
 		}
-		return h;
+		return static_cast<size_t>(h);
 	}
 };
